Task 3: Country with most phone records

+++ include/phone_record.h
@@ -1,12 +1,10 @@
 #ifndef PHONE_RECORD_H
 #define PHONE_RECORD_H
 
-#include <stdint.h>
-
 typedef struct {
-    uint64_t phone_number;
-    uint16_t country_code;
-    uint16_t year_created;
+    unsigned int phone_number;
+    unsigned short country_code;
+    unsigned short year_created;
     char *full_name;
 } phone_record_t;
